apps/server/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.0.1
openai==2.15.0
aiosqlite==0.22.1
python-multipart==0.0.20
